+++ CHANGELOG.md
@@ -16,7 +16,8 @@ All notable changes to this project will be documented in this file.
 ## [3.0.0] - 2017-11-23
 ## Changed
  - Add unit test against business logic on fleet.
- - Update unit test on flight due to the change of data structure. 
+ - Update unit test on flight due to the change of data structure.
+ - Add checking for sufficient budget on purchase.
 
 ## [3.0.0] - 2017-11-22
 ## Changed